Use bookmark if no input file(s) given
