+++ src/app/cli/cmd_print.go
@@ -9,14 +9,14 @@ import (
 
 type Print struct {
 	FilterArgs
-	MultipleFilesArgs
+	InputFilesArgs
 	Sort bool `short:"s" name:"sort" help:"Sort output by date (from oldest to latest)"`
 }
 
 func (args *Print) Run(ctx app.Context) error {
 	rs, err := ctx.RetrieveRecords(args.File...)
 	if err != nil {
-		return prettifyError(err)
+		return err
 	}
 	rs = service.FindFilter(rs, args.FilterArgs.toFilter())
 	if args.Sort {
